pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "anki_incontext"
authors = [{ name = "Abdo", email = "abdo@abdnh.net" }]
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
dynamic = ["version"]
dependencies = [
    "ankiutils",
    "pycountry>=24.6.1",
    "sentry-sdk>=1.14.0",
    "skell-downloader",
    "tdk",
]

[tool.setuptools]
package-dir = { "anki_incontext" = "src" }

[tool.uv.sources]
ankiscripts = { git = "https://github.com/abdnh/ankiscripts", branch = "master" }
ankiutils = { git = "https://github.com/abdnh/ankiutils", branch = "master" }
skell-downloader = { git = "https://github.com/abdnh/skell-downloader", branch = "master" }
tdk = { git = "https://github.com/abdnh/tdk", branch = "master" }

[dependency-groups]
dev = [
    "ankiscripts[forms]",
    "ankiutils",
    "aqt[qt]>=25.2.4",
    "mypy>=1.15.0",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "ruff>=0.11.7",
    "sentry-sdk>=1.14.0",
    "types-requests",
]

[project.urls]
Repository = "https://github.com/abdnh/anki-incontext"

[tool.ruff]
extend-exclude = [
    ".venv",
    "ankidata",
    "vendor",
    "forms",
    "user_files",
    "build",
    "node_modules",
]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "PL", "TRY"]
ignore = ["PLW0603", "PLR2004"]

[tool.pytest.ini_options]
addopts = "--cov=src --cov-branch"
testpaths = ["tests"]

[tool.coverage.run]
omit = ["src/forms/*", "src/vendor/*"]

[tool.mypy]
python_version = "3.9"
pretty = true
strict_optional = false
show_error_codes = true
disallow_untyped_decorators = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unused_ignores = true
check_untyped_defs = true
disallow_untyped_defs = true
strict_equality = true
exclude = ["venv", "ankidata", "vendor", "user_files", "build"]
modules = ["src", "tests"]

[[tool.mypy.overrides]]
module = "src.forms.*"
ignore_errors = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "src.vendor.*"
ignore_missing_imports = true
ignore_errors = true

[[tool.mypy.overrides]]
module = "bs4.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pycountry.*"
ignore_missing_imports = true
